moved login components into pages folder
